# model-service/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimExtractionRequest": {
        "description": "Request body for the claim extraction endpoint.",
        "properties": {
          "annotations": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Annotation data for context",
            "title": "Annotations"
          },
          "extraction_strategy": {
            "default": "sentence-based",
            "description": "Strategy for extracting claims",
            "enum": [
              "sentence-based",
              "semantic-units",
              "hierarchical"
            ],
            "title": "Extraction Strategy",
            "type": "string"
          },
          "max_claims": {
            "default": 50,
            "description": "Maximum number of claims to extract",
            "maximum": 200,
            "minimum": 1,
            "title": "Max Claims",
            "type": "integer"
          },
          "min_confidence": {
            "default": 0.5,
            "description": "Minimum confidence threshold for claims",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Confidence",
            "type": "number"
          },
          "ontology_glosses": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Map of type ID to gloss text",
            "title": "Ontology Glosses"
          },
          "ontology_types": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Ontology type definitions for context",
            "title": "Ontology Types"
          },
          "sentences": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Pre-split sentences (optional)",
            "title": "Sentences"
          },
          "summary_id": {
            "description": "Unique identifier for the summary",
            "minLength": 1,
            "title": "Summary Id",
            "type": "string"
          },
          "summary_text": {
            "description": "Full summary text to extract claims from",
            "minLength": 1,
            "title": "Summary Text",
            "type": "string"
          }
        },
        "required": [
          "summary_id",
          "summary_text"
        ],
        "title": "ClaimExtractionRequest",
        "type": "object"
      },
      "ClaimExtractionResponse": {
        "description": "Response body for the claim extraction endpoint.",
        "properties": {
          "claims": {
            "description": "Extracted claims",
            "items": {
              "$ref": "#/components/schemas/ExtractedClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "model_used": {
            "description": "LLM model used for extraction",
            "minLength": 1,
            "title": "Model Used",
            "type": "string"
          },
          "processing_time": {
            "description": "Processing time in seconds",
            "minimum": 0.0,
            "title": "Processing Time",
            "type": "number"
          },
          "summary_id": {
            "description": "Summary identifier",
            "minLength": 1,
            "title": "Summary Id",
            "type": "string"
          }
        },
        "required": [
          "summary_id",
          "model_used",
          "processing_time"
        ],
        "title": "ClaimExtractionResponse",
        "type": "object"
      },
...
      "DetectionRequest": {
        "description": "Request body for the object detection endpoint.",
        "properties": {
          "confidence_threshold": {
            "default": 0.3,
            "description": "Minimum confidence for detections",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Threshold",
            "type": "number"
          },
          "enable_tracking": {
            "default": true,
            "description": "Whether to enable object tracking",
            "title": "Enable Tracking",
            "type": "boolean"
          },
          "frame_numbers": {
            "description": "Specific frames to process",
            "items": {
              "type": "integer"
            },
            "title": "Frame Numbers",
            "type": "array"
          },
          "query": {
            "description": "Text query describing objects to detect",
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "video_id": {
            "description": "Unique identifier for the video",
            "minLength": 1,
            "title": "Video Id",
            "type": "string"
          },
          "video_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Optional full path to video file",
            "title": "Video Path"
          }
        },
        "required": [
          "video_id",
          "query"
        ],
        "title": "DetectionRequest",
        "type": "object"
      },
      "DetectionResponse": {
        "description": "Response body for the object detection endpoint.",
        "properties": {
          "frames": {
            "description": "Frames with detections",
            "items": {
              "$ref": "#/components/schemas/FrameDetections"
            },
            "title": "Frames",
            "type": "array"
          },
          "id": {
            "description": "Unique identifier for this detection job",
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "processing_time": {
            "description": "Processing time in seconds",
            "minimum": 0.0,
            "title": "Processing Time",
            "type": "number"
          },
          "query": {
            "description": "Query that was used",
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "total_detections": {
            "description": "Total detections across all frames",
            "minimum": 0,
            "title": "Total Detections",
            "type": "integer"
          },
          "video_id": {
            "description": "Video identifier",
            "minLength": 1,
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "video_id",
          "query",
          "total_detections",
          "processing_time"
        ],
        "title": "DetectionResponse",
        "type": "object"
      },
...
      "SummarySynthesisRequest": {
        "description": "Request body for the summary synthesis endpoint.",
        "properties": {
          "claim_relations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ClaimRelationship"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Relationships between claims",
            "title": "Claim Relations"
          },
          "claim_sources": {
            "description": "Claim hierarchies from one or more sources",
            "items": {
              "$ref": "#/components/schemas/ClaimSource"
            },
            "minItems": 1,
            "title": "Claim Sources",
            "type": "array"
          },
          "include_citations": {
            "default": false,
            "description": "Include inline citations to source claims",
            "title": "Include Citations",
            "type": "boolean"
          },
          "include_conflicts": {
            "default": true,
            "description": "Explicitly mention informational conflicts",
            "title": "Include Conflicts",
            "type": "boolean"
          },
          "max_length": {
            "default": 500,
            "description": "Maximum summary length in words",
            "maximum": 2000,
            "minimum": 100,
            "title": "Max Length",
            "type": "integer"
          },
          "ontology_context": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Ontology types and glosses for references",
            "title": "Ontology Context"
          },
          "persona_context": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Persona information for perspective",
            "title": "Persona Context"
          },
          "summary_id": {
            "description": "Target summary identifier",
            "minLength": 1,
            "title": "Summary Id",
            "type": "string"
          },
          "synthesis_strategy": {
            "default": "hierarchical",
            "description": "Strategy for organizing summary",
            "enum": [
              "hierarchical",
              "chronological",
              "narrative",
              "analytical"
            ],
            "title": "Synthesis Strategy",
            "type": "string"
          }
        },
        "required": [
          "summary_id",
          "claim_sources"
        ],
        "title": "SummarySynthesisRequest",
        "type": "object"
      },
      "SummarySynthesisResponse": {
        "description": "Response body for the summary synthesis endpoint.",
        "properties": {
          "claims_used": {
            "description": "Total claims synthesized",
            "minimum": 0,
            "title": "Claims Used",
            "type": "integer"
          },
          "model_used": {
            "description": "LLM model used for synthesis",
            "minLength": 1,
            "title": "Model Used",
            "type": "string"
          },
          "processing_time": {
            "description": "Processing time in seconds",
            "minimum": 0.0,
            "title": "Processing Time",
            "type": "number"
          },
          "summary_gloss": {
            "description": "Generated summary as GlossItem array with references",
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Summary Gloss",
            "type": "array"
          },
          "summary_id": {
            "description": "Summary identifier",
            "minLength": 1,
            "title": "Summary Id",
            "type": "string"
          },
          "synthesis_metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "description": "Metadata about synthesis",
            "title": "Synthesis Metadata",
            "type": "object"
          },
          "thinking": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThinkingTrace"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Optional reasoning trace from a thinking-capable model"
          }
        },
        "required": [
          "summary_id",
          "model_used",
          "processing_time",
          "claims_used"
        ],
        "title": "SummarySynthesisResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Cross-service contract for the request/response shapes the FOVEA model-service produces and the server consumes. Generated without importing the ML stack by scripts/gen_contract_spec.py; do not edit by hand.",
    "title": "FOVEA Model Service Contract",
    "version": "0.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/detection/detect": {
      "post": {
        "operationId": "post_/api/detection/detect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionResponse"
                }
              }
            },
            "description": "Successful response."
          }
        },
        "summary": "Detect objects in video frames."
      }
    },
    "/api/extract-claims": {
      "post": {
        "operationId": "post_/api/extract-claims",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimExtractionResponse"
                }
              }
            },
            "description": "Successful response."
          }
        },
        "summary": "Extract claims from a generated summary."
      }
    },
...
    "/api/synthesize-summary": {
      "post": {
        "operationId": "post_/api/synthesize-summary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarySynthesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummarySynthesisResponse"
                }
              }
            },
            "description": "Successful response."
          }
        },
        "summary": "Synthesize a summary gloss from claim sources."
      }
    }
  }
}